+++ index.html
@@ -451,7 +451,7 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
+    <script src="./js/main.js" type="module"></script>
     <script src="./js/functions.js"></script>
-    <script src="./js/main.js"></script>
   </body>
 </html>
