+++ js/util.js
@@ -1,0 +1,13 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+const getArrayFrom = (length, fnc) => Array.from({ length: length }, (_, index) => fnc(index + 1));
+
+export { getRandomInteger, getRandomArrayElement, getArrayFrom };
